flashcard-api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.1
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
PyPDF2==3.0.1
python-docx==0.8.11
